feat(atlas): S5 W1c — one Atlas database, the contract's five schemas in one file
A1 §5, "One Atlas DuckDB, multiple projection families", declares ONE physical
file (`atlas.duckdb`) carrying five logical schemas — meta, ops, source, git,
context — and decision A1-02's rationale is "DuckDB is already installed/
daemon-owned; schemas provide separation without more databases". S3 shipped
two files and the A1a register recorded that as a *ratified* deviation; no
owner ruling ratified it. The owner correction of 2026-08-29
(knowledge/rulings/owner-rulings/one-atlas-database-2026-08-29.md) settled that
the CODE converges to the contract. J4 — explicit owner decision; J5 for every
place the shipped code and A1 §5 disagreed, where the contract won.

What moved
- `src/runtime/analytics.rs` is deleted; its `ops.*` DDL, fold, canned queries
  and graph view now live in `src/runtime/atlas/db.rs`, which declares all five
  namespaces. `sergeant.duckdb`, `DUCKDB_FILE` and `duckdb_path` are gone.
  R2 — reuse the module that already owned a database rather than build a
  third thing; the merge is into ONE file because the one-owner invariant is
  about which module may reach the driver.
- `PROJECTIONS_DIR`/`projections_dir` moved to `runtime/startup.rs`, their only
  remaining consumer (the FloorState cache). `floor-state.json` did NOT move —
  relocating an unrelated cache is scope this wave was not given (R1).

The rebuild discipline survived as scope, not as separate files
- `Analytics::begin_rebuild` no longer deletes a file. It opens `atlas.duckdb`
  and runs `DROP SCHEMA IF EXISTS ops CASCADE` + recreate, which has exactly
  the reach `remove_file` had while `ops` had a file to itself. meta/source/
  git/context are untouched by a refold (F1).

The affordance changed meaning and is restated, not silently changed
- Deleting the projection file used to be lossless. Deleting `atlas.duckdb`
  still rebuilds every `ops` row from the journal AND discards every persisted
  source generation, which must be re-scanned. Both halves are MEASURED by
  `w1c_one_atlas_database::deleting_atlas_duckdb_rebuilds_ops_and_loses_source_facts`,
  not merely written down. `docs/concepts/atlas-and-knowledge.md` ("Two stores,
  two rebuild disciplines" -> "One store, two rebuild disciplines"), both
  module docs, `m5_projections`'s acceptance-1 doc and `sgt doctor`'s
  projection remedy all say what is now true; the remedy names a RESTART as
  the supported rebuild and warns against deleting the file to force one.

The join is proven, not assumed (owner ruling decision 4)
- `Analytics::work_overlay_generations` runs ONE statement joining
  `ops.work` to `source.generations`; `WORK_GENERATION_JOIN_SQL` is public so
  the test can assert what it is as well as what it returns.
  `w1c_one_atlas_database::one_statement_joins_ops_work_identity_to_source_generations`
  asserts a row carrying columns from both schemas, that a Work with no
  overlay is NOT joined in, and that the statement contains no ATTACH. No
  production caller (same posture as `Analytics::table_rows`) — A2's retrieval
  path is out of scope.

The two one-owner tests collapse into one — why the deleted test is correct
- `m5_projections::t2_the_duckdb_file_has_exactly_one_owner` and
  `x1_atlas_substrate::atlas_database_has_exactly_one_owner` existed as a PAIR
  only because there were two databases with two owning files, and both suites
  said in terms that they must never become a union rule ("either of these
  files may open a database"), which passes just as happily once one owner has
  grown into the other's territory. That premise no longer exists: one
  database, one owning module, one invariant. `x1`'s test now scans the WHOLE
  of `src/` with no exempted tree and no allowed-owners list — the STRONGER
  form, not a merged weaker one — and carries `t2`'s CLI half. A note where
  `t2` stood records the reasoning rather than leaving a silent deletion.

One file is one DuckDB instance (found by measurement, not assumed)
- Two `Connection::open` calls on the same path produce two INDEPENDENT
  database instances: they do not see each other's writes and the last close
  silently wins (probed directly in this lane before the design was fixed).
  Harmless while `ops` had its own file; data loss now. `Analytics::atlas()`
  hands out an `AtlasDb` over `Connection::try_clone` — same instance — and
  the daemon and API derive their Atlas handle from the open projection
  instead of calling `AtlasDb::open`, which now documents that it is for a
  process holding no other handle. It hands out an `AtlasDb`, never a
  `Connection`, so the one-owner invariant is untouched. It also skips
  `HARDENING_DDL` (instance-global and locked by the first open, so re-issuing
  it is refused) and VERIFIES F4's posture through `hardening()` instead.

Consequences reported rather than absorbed
- The daemon now creates `atlas.duckdb` on every start, because `ops` lives in
  it. So "this host has indexed nothing" can no longer be read off file
  absence, and `with_atlas`/`with_existing_atlas_write` read it off confirmed
  generations instead. The external answer is unchanged — same shape, same
  `detail` wording, same `atlas.present: false` — because the question it
  answers is unchanged. `w1b_overlay_lifecycle_trigger`'s R1 limit test is
  renamed `a_work_on_an_unindexed_estate_gains_no_atlas_evidence` and asserts
  the evidence, with a note saying why the file assertion is gone; the same
  restatement is made in `api::tests::an_unscanned_host_answers_without_writing_evidence`.
- `x1`'s token scan also matches lowercase prose spelling the file name, so
  documentation outside the owner says "Atlas's database file"/`ATLAS_DB_FILE`.
  An exception list would be how a scan stops seeing what it was written to
  see; the prose moved instead.

Register and evidence
- `x5_a1a_acceptance` item 1 is `met`, "ratified" struck, checks re-pointed at
  the five-namespace test and the new join test; the doc table row matches.
- Item 13's SQL-hole tripwire gains the four holes the moved ops code brings,
  each filled from `TABLES`/`MUTABLE_TABLES` — `&'static` constants in the same
  file — with `table_rows` validating a caller's name against `TABLES` before
  any formatting. Every other boundary test stays green unchanged.
- `w1c_one_atlas_database` is wired into `scripts/coverage/c2-suites.sh` at
  birth (the #231 lesson).

Deviation from the brief, stated rather than hidden: the brief said
`x2_knowledge_sources::source_facts_survive_a_real_daemon_restart` stays green
UNCHANGED. Its persistence half is untouched and green, but its *contrast*
half deleted `projections/` to demonstrate that the ops database was
disposable — and after the merge that directory holds only the FloorState
cache. The contrast is restated as what the daemon itself does (a plain
restart drops and refolds `ops`, source facts survive). Where the brief and the
contract disagree the contract wins (J5).

Full suite: 2280 tests, all green (one known-flaky `m9_watch` timing test
failed once under load and passes on re-run and in the prior full pass).
