+++ docs/concepts/atlas-and-knowledge.md
@@ -1,18 +1,22 @@
 # Atlas and knowledge sources
 
-Atlas is the daemon-owned store of what Sergeant has *learned about the world your estate works in*: the files under a declared knowledge path, the content of a repository at the commit an admission pinned, the structure a grammar found inside those files, and the datasets that stay where they are and get read in place. It is a second analytical database — `<data-dir>/atlas/atlas.duckdb` — deliberately separate from the operations projection that answers questions about Work.
+Atlas is the daemon-owned store of what Sergeant has *learned about the world your estate works in*: the files under a declared knowledge path, the content of a repository at the commit an admission pinned, the structure a grammar found inside those files, and the datasets that stay where they are and get read in place. It lives in the estate's one analytical database — `<data-dir>/atlas/atlas.duckdb` — alongside the operations projection that answers questions about Work, in its own schemas.
 
 Everything in Atlas is **derived evidence**. The authority is elsewhere and stays elsewhere: the journal for what happened, Git for repository content, and the operator's own bytes for everything read off a filesystem. Atlas holds a representation of those bytes that can be queried; it never holds the only copy of anything, and it never becomes the thing it describes.
 
-## Two stores, two rebuild disciplines
+## One store, two rebuild disciplines
 
-Sergeant keeps two DuckDB files and they do not share a rebuild story. Confusing them is the single most likely way to lose derived evidence, so the difference is worth stating plainly.
+Sergeant keeps **one** DuckDB file, `atlas.duckdb`, with five schemas: `meta`, `ops`, `source`, `git` and `context`. One database is what makes a question that spans two of them — "which source generations belong to this Work?" — an ordinary join instead of a federation across separate stores.
 
-The **operations projection** (`sergeant.duckdb`, holding the `ops` schema) is disposable. The daemon deletes it and re-folds it from the journal on every start. "Delete this file" and "restart the daemon" are the same operation, and nothing may come to depend on state that lives only there.
+What that one file does *not* have is one rebuild story, and confusing the two is the single most likely way to lose derived evidence.
 
-**Atlas persists.** Its `source` tables and its coverage rows are not a function of the journal — no replay reproduces them, because they are derived from source *bytes* plus the identity of the extractor that read those bytes. Deleting `atlas.duckdb` is therefore not free and not a no-op: it is re-derivable only by re-scanning every declared source, which costs whatever those sources cost to read, and until that happens coverage will correctly report that the derived evidence is gone.
+The **operations projection** (the `ops` schema) is disposable. It is a pure fold of the journal, and the daemon drops the whole `ops` schema and re-folds it on every start. "Restart the daemon" is the only rebuild there is, and nothing may come to depend on state that lives only there.
 
-What keeps the journal authoritative across the split is a summary event. Each completed scan journals exactly one compact `source.scanned` record — the source, the generation, its content key, the counts by coverage status, and which extractors ran — and never a per-file event stream. The journal carries the trail; Atlas carries the detail. A crash between the two is a resolved case rather than an open one: a scan's rows land as one provisional generation, the summary is journaled, and a second transaction confirms it. No read path can see an unconfirmed generation, and the next daemon start resolves any a crash left behind — by asking the journal, because the two crash windows have opposite correct answers and the database cannot tell them apart on its own. A generation whose `source.scanned` summary never reached the journal is **evicted**: that scan never completed, and the eviction leaves an explicit `generation_evicted` coverage row naming the window it closed. A generation whose summary *is* durable is **confirmed** instead, exactly as the live path would have confirmed it — the journal already vouches for that scan and has already broadcast it, so discarding the rows would leave the trail claiming a scan the store had thrown away. A start that followed no crash finds nothing provisional and does neither.
+**Everything else persists.** The `source` tables and the coverage rows are not a function of the journal — no replay reproduces them, because they are derived from source *bytes* plus the identity of the extractor that read those bytes.
+
+So: **restarting the daemon rebuilds `ops` and costs nothing. Deleting `atlas.duckdb` is a different act, and it is not free.** Deleting the file still brings every `ops` row back from the journal on the next start — and it discards every persisted source generation, which is re-derivable only by re-scanning every declared source, at whatever those sources cost to read. Until that happens `sgt intelligence status` and `sgt doctor` will correctly report that this host has indexed nothing. If what you actually want is a fresh operations projection, restart the daemon; there is no reason to delete the file.
+
+What keeps the journal authoritative across that split is a summary event. Each completed scan journals exactly one compact `source.scanned` record — the source, the generation, its content key, the counts by coverage status, and which extractors ran — and never a per-file event stream. The journal carries the trail; Atlas carries the detail. A crash between the two is a resolved case rather than an open one: a scan's rows land as one provisional generation, the summary is journaled, and a second transaction confirms it. No read path can see an unconfirmed generation, and the next daemon start resolves any a crash left behind — by asking the journal, because the two crash windows have opposite correct answers and the database cannot tell them apart on its own. A generation whose `source.scanned` summary never reached the journal is **evicted**: that scan never completed, and the eviction leaves an explicit `generation_evicted` coverage row naming the window it closed. A generation whose summary *is* durable is **confirmed** instead, exactly as the live path would have confirmed it — the journal already vouches for that scan and has already broadcast it, so discarding the rows would leave the trail claiming a scan the store had thrown away. A start that followed no crash finds nothing provisional and does neither.
 
 ## Sources are evidence; repositories are authority
