+++ src/runtime/mod.rs
@@ -1,7 +1,6 @@
 //! Runtime: journal, projections, atlas, git, surfaces, integrity, repository
 //! locking, git admission preflight, routing, engine, recovery, sweep.
 
-pub mod analytics;
 pub mod atlas;
 pub mod blob;
 pub mod engine;
